% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-extdata.R
\name{make_ext_chas_datadictionary}
\alias{make_ext_chas_datadictionary}
\title{Download External Data: 'CHAS data dictionary 11-15.xlsx'}
\usage{
make_ext_chas_datadictionary()
}
\value{
A \code{data.frame} object
}
\description{
Download the external dataset "CHAS data dictionary 11-15.xlsx" from osf.io \cr
This is the data dictionary for CHAS' housing cost burden data from 2011-2015 at
the Census Place geographic scale.
}
