% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-extdata.R
\name{make_ext_chas_housingcostburden}
\alias{make_ext_chas_housingcostburden}
\title{Download External Data: Table-8.csv}
\usage{
make_ext_chas_housingcostburden()
}
\value{
A \code{data.frame} object
}
\description{
Download the external dataset Table-8.csv from osf.io \cr
This is the CHAS data on housing cost burden from 2011-2015 at
the Census Place geographic scale.
}
