% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans-external-data.R
\docType{data}
\name{plan_external_data}
\alias{plan_external_data}
\title{Drake Plan: External Data}
\format{A \code{\link[drake]{drake_plan}} object of class \code{tbl_df} that builds
the following \code{\link[drake]{target}}s:
\describe{

\item{\code{• ext_chas_housingcostburden}:}{Housing cost burden data (2011-2015) from CHAS}
\item{\code{• ext_chas_datadictionary}:}{Data dictionary for the CHAS data}

}}
\usage{
plan_external_data
}
\description{
A \code{\link[drake]{drake_plan}} that loads several external datasets.
}
\examples{

# Print the plan

print(plan_external_data)


# Make the plan, load a target, print the target

\dontrun{

make(ext_chas_housingcostburden)

readd(ext_chas_housingcostburden)
}
}
\seealso{
\code{\link{data_exteral_datasets_list}}
}
\keyword{drake,}
\keyword{plan}
